Move 'times_to_microseconds' and 'microseconds_to_times' from utils to tmoc, remove utils
