+++ mocpy/tests/test_tmoc.py
@@ -3,9 +3,35 @@ import pytest
 from astropy.time import Time
 
 # from astropy.io import ascii
-from ..tmoc import TimeMOC
+from ..tmoc import *
 
 import numpy as np
+
+
+def test_time_to_microsec_1():
+    """Test of the time conversion from `astropy.time` with format isot to microseconds."""
+    # t = Time([['1998-01-01', '1999-01-01']], format="iso", scale="tdb")
+
+    t = Time("1999-01-01T00:00:00.123456789", format="isot", scale="tdb")
+    # t = Time('0000-01-01T00:00:00.123456789', format='isot', scale='tdb')
+    # print(t.jd * utils.DAY_MICRO_SEC)
+    us = times_to_microseconds(t)
+    jd = microseconds_to_times(us)
+    assert us == 211781908800123456
+    assert jd.jd == t.jd
+
+
+def test_time_to_microsec_2():
+    """Test of the time conversion from `astropy.time` with format iso to microseconds."""
+    t = Time([["1998-01-01", "1999-01-01"]], format="iso", scale="tdb")
+    us1 = times_to_microseconds(t)
+    us2 = np.asarray(t.jd * 86400000000, dtype=np.uint64)
+    jd1 = microseconds_to_times(us1)
+    jd2 = microseconds_to_times(us2)
+    print(us1)
+    print(us2)
+    assert (us1 == us2).all()
+    assert (jd1.jd == jd2.jd).all()
 
 
 def test_complement():
