+++ mocpy/tmoc/__init__.py
@@ -1,5 +1,3 @@
-from .tmoc import TimeMOC
+from .tmoc import TimeMOC, times_to_microseconds, microseconds_to_times
 
-__all__ = [
-    "TimeMOC",
-]
+__all__ = ["TimeMOC", "times_to_microseconds", "microseconds_to_times"]
